Fixed a problem when dealing with GBK and Big5 double byte characters
In -feedBytes:length:connection: of YLTerminal.mm, we process the input
bytes one by one, if we found that byte is 0x9B (Control Sequence
Introducer), we will enter TP_CONTROL state, but the truth is, we may
met 0x9B as the lower byte in GBK too! So, we need another state to
track whether we are waiting for the lower byte of GBK/Big5, thus I
added TP_NEXT_BYTE in this commit, in TP_NEXT_BYTE state, it will
simply save that lower byte regardless it's special or not to avoid
the above problem.


git-svn-id: https://nally.googlecode.com/svn/trunk@82 db928f3c-c32d-11dd-ba5c-d3b7c2b3b534