Make the tracking area larger
Now the tracking area expands to the entire the image view. That
created a problem: when the mouse cursor is already in the that
view then it's created, you have to move the cursor out of it
then move it back in to see the download indicator. That's kind
of stupid. We fixed it here by first enable the mouseMoved event
to track the cursor move at first, then disable it as soon as it
moved the first time we stop tracking mouseMoved, because we only
need to track mouseEntered and mouseExited after that.

git-svn-id: https://nally.googlecode.com/svn/trunk@228 db928f3c-c32d-11dd-ba5c-d3b7c2b3b534